webserver/requirements.txt
Django==1.11.5
djangorestframework==3.6.4
djangorestframework-jwt==1.11.0
psycopg2==2.7.3.1
PyJWT==1.5.3
pytz==2017.2
